Bugs fixed

+++ app/src/main/java/com/mid/bilweatherapp/json/WeatherResponse.kt
@@ -108,14 +108,14 @@ data class Hour (
     @SerializedName("condition"      ) var condition    : Condition? = Condition(),
     @SerializedName("wind_mph"       ) var windMph      : Double?    = null,
     @SerializedName("wind_kph"       ) var windKph      : Double?    = null,
-    @SerializedName("wind_degree"    ) var windDegree   : Int?       = null,
+    @SerializedName("wind_degree"    ) var windDegree   : Double?       = null,
     @SerializedName("wind_dir"       ) var windDir      : String?    = null,
-    @SerializedName("pressure_mb"    ) var pressureMb   : Int?       = null,
+    @SerializedName("pressure_mb"    ) var pressureMb   : Double?       = null,
     @SerializedName("pressure_in"    ) var pressureIn   : Double?    = null,
     @SerializedName("precip_mm"      ) var precipMm     : Double?    = null,
     @SerializedName("precip_in"      ) var precipIn     : Double?    = null,
-    @SerializedName("snow_cm"        ) var snowCm       : Int?       = null,
-    @SerializedName("humidity"       ) var humidity     : Int?       = null,
+    @SerializedName("snow_cm"        ) var snowCm       : Double?       = null,
+    @SerializedName("humidity"       ) var humidity     : Double?       = null,
     @SerializedName("cloud"          ) var cloud        : Int?       = null,
     @SerializedName("feelslike_c"    ) var feelslikeC   : Double?    = null,
     @SerializedName("feelslike_f"    ) var feelslikeF   : Double?    = null,
@@ -129,11 +129,11 @@ data class Hour (
     @SerializedName("chance_of_rain" ) var chanceOfRain : Int?       = null,
     @SerializedName("will_it_snow"   ) var willItSnow   : Int?       = null,
     @SerializedName("chance_of_snow" ) var chanceOfSnow : Int?       = null,
-    @SerializedName("vis_km"         ) var visKm        : Int?       = null,
-    @SerializedName("vis_miles"      ) var visMiles     : Int?       = null,
+    @SerializedName("vis_km"         ) var visKm        : Double?       = null,
+    @SerializedName("vis_miles"      ) var visMiles     : Double?       = null,
     @SerializedName("gust_mph"       ) var gustMph      : Double?    = null,
     @SerializedName("gust_kph"       ) var gustKph      : Double?    = null,
-    @SerializedName("uv"             ) var uv           : Int?       = null
+    @SerializedName("uv"             ) var uv           : Double?       = null
 
 )
